Get LERI data

This function searches for LERI data by date, returning a Raster* object.

get_leri(date, product, dir = tempdir(), overwrite = FALSE)

Arguments

date
An object of class Date or a character string formatted as are to be acquired. To specify a time interval or date range, date can be a vector of class Date such as produced by seq.Date.

product
A string that specifies which LERI product to get, e.g., "1 month", "3 month", "7 month", or "12 month" for their monthly data, or "8 day ac", or "8 day nac" for accumulated and non-accumulated (respectively) 8-day LERI products. Fractional timescales are not supported, and will be rounded to the nearest integer (e.g., "1.1 month" will be converted to "1 month").

dir
Directory to for downloaded LERI data. By default this will be a temporary directory. This should be a file path specified as a string.

overwrite
Boolean to indicate whether to overwrite LERI data that already exist locally in dir. Defaults to FALSE.

Value

A Raster* object containing LERI data. Each layer in this object corresponds to a unique LERI data file. In cases where one LERI file covers multiple dates that were provided in the date argument, the number of layers in the returned Raster* object may be less than the number of dates queried. For example, if a user requests the 8 day LERI product for two consecutive dates, both of those dates might be contained in the same 8 day LERI data product. The time intervals covered by each layer in the Raster* object returned by get_leri are available as layer names.

Details

The Landscape Evaporative Response Index (LERI) is available from 2000 to present, with a ~5 day lag to the current date. Products are available at multiple timescales. Monthly LERI data are available at the 1, 3, 7, and 12 month timescale for each month of the year. During the growing season (April through the end of October), there are two additional LERI products:

"Non-accumulated" 8-day data, which represent deviations in LERI within an eight day time frame (for example, the time period from May 01 to May 08).

"Accumulated" 8-day data, which represent deviations in LERI for the period from April 1 to the end of the eight day time frame (for example, the time period from April 01 to May 08).

The values in LERI rasters are percentiles for estimates of actual evapotranspiration computed by the operational Simplified Surface Energy Balance (SEEBop) model, which uses remotely sensed MODIS thermal imagery with climatological reference evapotranspiration as descriped in Senay et al., 2013.

More information about how the monthly, non-accumulated 8-day, and accumulated 8-day data differ can be found at https://www.esrl.noaa.gov/psd/leri/.

References

Senay, Gabriel B., Stefanie Bohms, Ramesh K. Singh, Prasanna H. Gowda, Naga M. Velpuri, Henok Alemu, James P. Verdin, 2013b. Operational Evapotranspiration Mapping Using Remote Sensing and Weather Datasets: A New Parameterization for the SSEB Approach. Journal of the American Water Resources Association (JAWRA). 49(3):577-591. http://onlinelibrary.wiley.com/doi/10.1111/jawr.12057/full

Examples

# \donttest{
# this may take a while, as it downloads ~72 MB
get_leri(date = "2018-01-01", product = "1 month")

#> class      : RasterLayer 
#> dimensions : 2844, 6617, 18818748  (nrow, ncol, ncell)
#> resolution : 0.009, 0.009  (x, y)
#> extent     : -126.009, -66.456, 23.949, 49.545  (xmin, xmax, ymin, ymax)
#> crs        : +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0 
#> source     : memory
#> names      : LERI_01mn_20180101.nc 
#> values     : 3.278688, 96.72131  (min, max)
#> 

# multi-day example that returns a two-layer raster
r <- get_leri(date = c("2018-08-12", "2018-08-13"), product = "8 day nac")
names(r) # has two elements (Aug05-Aug12, and Aug13-Aug20)

#> [1] "LERI_8day.nac_Aug05.Aug12_2018.nc" "LERI_8day.nac_Aug13.Aug20_2018.nc"

# }